backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-dotenv>=1.0,<2
faker>=33,<35
chromadb>=0.6,<1
google-genai>=1.0,<2
